+++ R/svyglmnb.R
@@ -96,7 +96,6 @@ svyglm.nb <- function(formula, design, ...) {
   attr(svyfit, "family") <- fam
   attr(svyfit, "nb.theta") <- mod[["theta"]]
   attr(svyfit, "nb.theta.se") <- mod[["SE.theta"]]
-  attr(svyfit, "response.median") <- svy_md(stats::model.frame(svyfit)[[1]], design = design)
 
   svyfit
 }